MobileNet-V3/requirements.txt
torch==1.9.0
torchvision==0.10.0
torchsummary==1.5.1
Pillow==8.2.0
ipython==7.23.1
jupyter==1.0.0
